+++ code/cell.module/src/Builder/index.ts
@@ -1,8 +1,1 @@
-import { Builder as B } from './types';
-import { chain } from './builder';
-
-export const Builder: B = {
-  chain: chain as any,
-};
-
-// export * from './Builder';
+export * from './builder';
